Update 像素硬边缘放大.cs
修复上次更新中，奇数长宽素材的平移/裁切中心没有量化到整数的问题。

+++ 像素硬边缘放大.cs
@@ -205,7 +205,7 @@ namespace Test_Script1
                             vKeyframes[0].Type = VideoKeyframeType.Linear;
                             VideoMotionBounds bounds = new VideoMotionBounds(new VideoMotionVertex(0f, 0f), new VideoMotionVertex((float)scrWidth, 0f), new VideoMotionVertex((float)scrWidth, (float)scrHeight), new VideoMotionVertex(0f, (float)scrHeight));
                             vKeyframes[0].Bounds = bounds;
-                            vKeyframes[0].MoveBy(new VideoMotionVertex(dFullWidth / 2 - scrWidth / 2, dFullHeight / 2 - scrHeight / 2));
+                            vKeyframes[0].MoveBy(new VideoMotionVertex((int)dFullWidth / 2 - scrWidth / 2, (int)dFullHeight / 2 - scrHeight / 2));
 
                             // If the material's height (or width) is larger than the project's, add PositionOFX
                             if (dFullWidth > scrWidth || dFullHeight > scrHeight)
